requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
boto3==1.26.42
botocore==1.29.42
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.1.2
contourpy==1.0.6
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==2.0.0
Django==4.1.4
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-heroku==0.3.1
django-js-asset==2.0.0
djangorestframework==3.14.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.2
fonttools==4.38.0
fqdn==1.5.1
h11==0.14.0
idna==3.4
ipykernel==6.19.4
ipython==8.8.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.2
jupyter_server==2.0.6
jupyter_server_terminals==0.4.3
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.18.0
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mysql-connector-python==8.0.33
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
openpyxl==3.1.2
outcome==1.2.0
packaging==22.0
panda==0.3.1
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pipenv==2022.11.30
platformdirs==2.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
psycopg2==2.9.6
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.4
pytz==2022.7
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
s3transfer==0.6.0
selenium==4.11.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.8.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.6.3
tzdata==2022.7
undetected-chromedriver==3.5.2
uri-template==1.2.0
urllib3==1.26.13
validate-email==1.3
virtualenv==20.17.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
webcolors==1.12
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.4.2
websockets==11.0.3
whitenoise==6.5.0
wsproto==1.2.0
